fix: pass MouseEvent to onClick handler from onInputElementClick
We encountered an issue that was described here, https://github.com/react-dropzone/react-dropzone/issues/40, that describes attempting to load the same file twice (the data has changed). The use-case where we encountered this was a user uploaded a file, but the rejected it due to formatting issues. The file input doesn't discard the target state so uploaded the correct file moments later does not trigger any handles such as `onDrop`, etc....

Summary of changes:
  - Update that event is passed to `onClick` prop handler inside of
`onInputElementClick`
  - Update `'should invoke inputProps onClick if provided'` testcase to
verify that function is now called with a MouseEvent object