feat: adding single prop to override multiple
This allows an opt in shortcut prop of `single` to turn off the
default `multiple` prop.

Example:
```
// shortcut boolean single prop
<DropZone single>...</DropZone>
// versus old default true multiple prop
<DropZone multiple={false}>...</DropZone>
```

This works on the assumption that if a user is providing single, as
opposed to the default true value of multiple, they intend for the
component to work as `multiple={false}`.

If provided BOTH props with a true value, it seems reasonable that
since the default is already `multiple={true}`, the presence of
`single` is a signal that the user intends the component to be single.
Otherwise, there would be no way of discerning if `multiple === true`
is a product of the default value or the user explicitly including it.

Additionally, this would provide an avenue to eventually deprecate `multiple={false}` which seems like a missed opportunity to provide users a more
concise prop.